+++ pkg/services/authz/zanzana/dualwrite/reconciler.go
@@ -1,4 +1,4 @@
-package migrator
+package dualwrite
 
 import (
 	"context"
@@ -599,4 +599,21 @@ func fixedRoleTuplesCollector(store db.DB) TupleCollector {
 
 		return nil
 	}
+}
+
+func batch(count, batchSize int, eachFn func(start, end int) error) error {
+	for i := 0; i < count; {
+		end := i + batchSize
+		if end > count {
+			end = count
+		}
+
+		if err := eachFn(i, end); err != nil {
+			return err
+		}
+
+		i = end
+	}
+
+	return nil
 }
